Partitioned the button testing steps into their own PHP trait file.
